fix: further spell updates
I'mma be honest I don't know how this happened, and I don't feel like
looking through old commits. Somehow half of the Rank 5 and 6 spells
were different from my source. I have no idea if I used it or not, but
it feels like a ChatGippity thing, so I'm going to blame that.